requirements.txt
aiofiles==23.1.0
aiohttp==3.8.1
khl.py==0.3.15
leancloud==2.9.12
pandas==1.5.2
Pillow==9.4.0
psutil==5.9.4
requests==2.28.1
zhconv==1.4.3
pyg2plot==0.1.8